day9.cpp
Problem : Permutations of a String
You are given a string s. Your task is to generate and return all possible permutations of the characters in the string. A permutation is a rearrangement of the characters in the string, and each character must appear exactly once in every permutation. If there are duplicate characters in the string, the resulting permutations should also be unique (i.e., no repeated permutations).

Input :
A string s consisting of lowercase English letters. The length of the string n satisfies 1≤𝑛≤9.
Input: "abc"

Output :
An array of strings containing all unique permutations of the input string. The order of permutations in the output does not matter.
Output: ["abc", "acb", "bac", "bca", "cab", "cba"]